pyproject.toml
[project]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: C",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS"
]
dependencies = [
 "wcwidth>=0.2.8",
 "pythonnet>=3.0.5",
 "setuptools>=80.9.0",
]
description = "JSON formatter that produces highly readable but fairly compact output."
license = "MIT"
name = "fractured-json"
readme = "README.md"
requires-python = ">=3.11, <3.14"
version = "5.0.0post5"

[[project.authors]]
email = "python@figsandfudge.com"
name = "Jon Connell"

[dependency-groups]
dev = [
 "pytest-console-scripts",
 "pytest-cov",
 "requests>=2.32.5",
 "ruff",
 "beautifulsoup4>=4.14.3",
 "pytest-check>=2.6.2",
 "tox>=4.32.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
documentation = "https://github.com/masaccio/fractured-json-python/blob/main/README.md"
repository = "https://github.com/masaccio/fractured-json-python"

[project.scripts]
fractured-json = "fractured_json._fractured_json:main"

[tool.pytest.ini_options]
addopts = "--cov=src/fractured_json --cov-report=html --cov-report=term-missing:skip-covered --cov-context=test"

[tool.ruff]
fix = true
ignore = [ "T201",]
line-length = 100
select = [ "ALL",]
target-version = "py311"

[tool.tox.global]
isolated_build = true
envlist = ["py311", "py312", "py313"]

[tool.tox.env_run_base]
description = "Run test under {base_python}"
allowlist_externals = ["pytest"]
commands = [["pytest", "--quiet", "--no-cov", "--import-mode", "importlib"]]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
"src/fractured_json/__init__.pyi" = [ "ANN401", "PYI035",]
"src/fractured_json/generated/**" = [ "E501", "D100", "D104",]
"tests/**" = [ "PTH123", "PLR2004", "S101", "D100", "D103", "ANN201", "ANN001",]

[tool.coverage.run]
branch = true
omit = [ "src/fractured_json/generated/*.py",]

[tool.coverage.html]
directory = "coverage_html_report"
show_contexts = true

